DEV: Review fixes
