+++ app/assets/javascripts/discourse/app/lib/sidebar/user/community-section/admin-revamp-section-link.js
@@ -31,10 +31,9 @@ export default class AdminRevampSectionLink extends BaseSectionLink {
 
     return (
       this.currentUser.staff &&
-      this.currentUser.isInAnyGroups(
-        this.siteSettings.groupSettingArray(
-          "enable_experimental_admin_ui_groups"
-        )
+      this.siteSettings.userInAnyGroups(
+        "enable_experimental_admin_ui_groups",
+        this.currentUser
       )
     );
   }
